@@ -5,33 +5,32 @@
     "keywords": ["cms", "azuriom", "laravel"],
     "license": "GPL-3.0-or-later",
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^8.0",
         "ext-json": "*",
         "ext-zip": "*",
         "bacon/bacon-qr-code": "^2.0",
         "composer/semver": "^3.2",
-        "fruitcake/laravel-cors": "^2.0",
+        "fruitcake/laravel-cors": "^2.0.5",
         "funkjedi/composer-include-files": "^1.1",
         "google/recaptcha": "^1.2",
-        "guzzlehttp/guzzle": "^7.0.1",
-        "laravel/framework": "^8.75",
-        "laravel/socialite": "^5.2",
-        "laravel/tinker": "^2.5",
-        "laravel/ui": "^3.1",
+        "guzzlehttp/guzzle": "^7.2",
+        "laravel/framework": "^9.0",
+        "laravel/socialite": "^5.4",
+        "laravel/tinker": "^2.7",
+        "laravel/ui": "^3.4.2",
         "pragmarx/google2fa": "^8.0",
-        "symfony/filesystem": "^5.1.4",
+        "symfony/filesystem": "^6.0",
         "textalk/websocket": "^1.5",
         "thedudeguy/rcon": "^1.0",
-        "truckersmp/steam-socialite": "^1.3",
-        "xpaw/php-source-query-class": "^1.0"
+        "xpaw/php-source-query-class": "^2.1"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.5",
-        "facade/ignition": "^2.5",
+        "barryvdh/laravel-debugbar": "^3.6.6",
         "fakerphp/faker": "^1.9.1",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^5.10",
-        "phpunit/phpunit": "^9.5.10"
+        "nunomaduro/collision": "^6.0",
+        "phpunit/phpunit": "^9.5.10",
+        "spatie/laravel-ignition": "^1.0"
     },
     "autoload": {
         "psr-4": {
