@@ -97,49 +97,36 @@ class NavbarElement extends Model
 
     public function getLink()
     {
-        switch ($this->type) {
-            case 'home':
-                return route('home');
-            case 'link':
-                return $this->value;
-            case 'page':
-                return route('pages.show', $this->value);
-            case 'post':
-                return route('posts.show', $this->value);
-            case 'posts':
-                return route('posts.index');
-            case 'plugin':
-                return Route::has($this->value) ? route($this->value) : '#';
-            default:
-                return '#';
-        }
+        return match ($this->type) {
+            'home' => route('home'),
+            'link' => $this->value,
+            'page' => route('pages.show', $this->value),
+            'post' => route('posts.show', $this->value),
+            'posts' => route('posts.index'),
+            'plugin' => Route::has($this->value) ? route($this->value) : '#',
+            default => '#',
+        };
     }
 
     public function isCurrent()
     {
         $request = request();
 
-        switch ($this->type) {
-            case 'home':
-                return $request->routeIs('home');
-            case 'link':
-                return $request->is($this->value);
-            case 'page':
-                return $request->routeIs('pages.show') && $request->route('page.slug') === $this->value;
-            case 'post':
-                return $request->routeIs('posts.show') && $request->route('post.slug') === $this->value;
-            case 'posts':
-                return $request->routeIs('posts.*');
-            case 'plugin':
-                return $request->routeIs(Str::beforeLast($this->value, '.').'.*');
-            case 'dropdown':
-                return $this->elements
-                    ->contains(function (self $element) {
-                        return ! $element->isDropdown() && $element->isCurrent();
-                    });
-            default:
-                return false;
-        }
+        return match ($this->type) {
+            'home' => $request->routeIs('home'),
+            'link' => $request->is($this->value),
+            'page' => $request->routeIs('pages.show') && $request->route('page.slug') === $this->value,
+            'post' => $request->routeIs('posts.show') && $request->route('post.slug') === $this->value,
+            'posts' => $request->routeIs('posts.*'),
+            'plugin' => $request->routeIs(Str::beforeLast($this->value, '.').'.*'),
+            'dropdown' => $this->elements
+                ->contains(
+                    function (self $element) {
+                    return ! $element->isDropdown() && $element->isCurrent();
+                }
+                ),
+            default => false,
+        };
     }
 
     public function getNameAttribute(string $value)
