@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateActionLogsTable extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -18,6 +17,7 @@ class CreateActionLogsTable extends Migration
             $table->unsignedInteger('user_id');
             $table->string('action');
             $table->unsignedInteger('target_id')->nullable();
+            $table->text('data')->nullable();
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->cascadeOnDelete();
@@ -33,4 +33,4 @@ class CreateActionLogsTable extends Migration
     {
         Schema::dropIfExists('action_logs');
     }
-}
+};
