@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateServersTable extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -34,4 +33,4 @@ class CreateServersTable extends Migration
     {
         Schema::dropIfExists('servers');
     }
-}
+};
